requirements.txt
sentence-transformers==2.5.1
pandas==1.4.3
opencv-python-headless==3.4.14.53
matplotlib==3.5.3
seaborn==0.13.2
umap-learn==0.5.6
